# _repr_ and _html_repr_ show '... and additional rows' message