…ncated outputs

# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #1026

 # Rationale for this change
<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->
Current implementation of ```_repr_``` and ```_html_repr_``` has no indication for whether the output was truncated or not.

# What changes are included in this PR?
<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->
The changes override the default output for the ```_repr_``` and ```_html_repr_``` functions and, instead, check if the table has more than 10 rows. If it does, then it prints the first 10 rows along with a ```... and additional rows``` message.

A thing to note is that these changes flatten the batches of the dataframe in the ```_repr_ function``` (as already done for the ```_repr_html_ function```). This was done primarily since using ```.collect``` twice was causing major performance degradation. Batch flattening does this at 1/10 th of the time.

# Are there any user-facing changes?
<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->
The users shall now see the ```... and additional rows``` message along with the first 10 rows if the table has more than 10 rows.

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->