@Spaarsh 2025-03-03
This is the new output:
## For ```_repr_```
```
+---------+---------+
| letters | numbers |
+---------+---------+
| A       | 1       |
| B       | 2       |
| C       | 3       |
| D       | 4       |
| E       | 5       |
| F       | 6       |
| G       | 7       |
| H       | 8       |
| I       | 9       |
| J       | 10      |
+---------+---------+
and more...
```

## For ```_repr_html_```
```
<table border='1'>
<tr><th>letters</th><th>numbers</th></tr>
<tr><td>A</td><td>1</td></tr>
<tr><td>B</td><td>2</td></tr>
<tr><td>C</td><td>3</td></tr>
<tr><td>D</td><td>4</td></tr>
<tr><td>E</td><td>5</td></tr>
<tr><td>F</td><td>6</td></tr>
<tr><td>G</td><td>7</td></tr>
<tr><td>H</td><td>8</td></tr>
<tr><td>I</td><td>9</td></tr>
<tr><td>J</td><td>10</td></tr>
<tr><td colspan="100%">and more...</td></tr>
</table>
```

## Performance Changes
| Implementation | `_repr_` | `_html_repr_` |
|---------------|----------|---------------|
| Current       | 3.16ms   | 1.35ms       |
| New           | 5.5ms  | 4.5ms       |

*Note: I have manually ran the command multiple times and observed this. If required, I will run a script and produce the average for both cases*