@Spaarsh 2025-03-06
On second thoughts, since this change will lead to a performance impact (refer the last part of this [comment](https://github.com/apache/datafusion-python/pull/1041#issuecomment-2694647145)), would it be better if we keep it separate? Just for the sake of documentation and clarity down the lane?